Don't loop when importing data
We have tested and fixed 'big' data problems, server side tables and
map clustering

+++ django_project/sql_scripts/30_load_data.sql
@@ -742,7 +742,7 @@ INSERT INTO features.feature_attribute_value(
 ) SELECT
     ft.feature_uuid as feature_uuid,
     22 as attribute_id,
-    static_water_level as val_real,
+    static_water_level::float as val_real,
     1 as changeset_id
 FROM
     test_data.import_raw_data ird
